adding some boundary features.
I added some boundary detection and I can list what’s in view, but I
don’t know if I can adapt this to load *only* what’s in view from the
JSON file. I think this is going got have to be a feature that comes
from the database integration.

It’s still a neat trick, and I kind of like it. Might leave it there,
if I can figure out a way to click on a name in that list and make its
popup open up. That would be kind of neat.